refactor layer config
